Minor cleanups
* Move wrapper_handler tests to its own file
* Use jsonschema validation for wrapper request
* Move unit test path management to __init__.py
