+++ test/unit/__init__.py
@@ -1,0 +1,4 @@
+# make sure we can find the app code
+import sys, os
+my_path = os.path.dirname(os.path.abspath(__file__))
+sys.path.insert(0, my_path + '/../../financial_functions/')
