+++ test/unit/test_core.py
@@ -1,10 +1,5 @@
 import pytest
 
-# make sure we can find the app code
-import sys, os
-my_path = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, my_path + '/../../financial_functions/')
-
 from datetime import date
 import core as ff
